backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
pdfplumber
requests
python-dotenv
scikit-learn
transformers
torch
sentencepiece